fix: prevent ordering unlisted products
users were able to place orders for unlisted products through the cart. This commit addresses the issue by adding a new condition in the OrderController to restrict ordering unlisted products.

Changes Made:

Added a validation check in the OrderController to prevent users from ordering unlisted products.
Resolved the issue where unlisted products could be added to the cart and ordered.

#to-do - have to do add delivery charge